test(cli-e2e): separate collect and config E2E tests
